[Refactor #41] 업로드 API 오류수정 및 API 모듈 네이밍 변경
